Various ios specific fixes

+++ etw/specials.c
@@ -1788,8 +1788,8 @@ BOOL VideoPrefs(WORD button)
                 /* AC: if we are on screen, disable the scaling */
                 use_gfx_scaling = FALSE;
 
-                ResizeWindow(atol(resolutions[current_resolution]),
-                             atol(resolutions[current_resolution] + 4));
+                ResizeWindow(atoi(resolutions[current_resolution]),
+                             atoi(resolutions[current_resolution] + 4));
                 
                 /* AC: Since these buttons are restored later on, I think
                  * that we don't have to eliminate them. They are Scaling
@@ -1807,8 +1807,8 @@ BOOL VideoPrefs(WORD button)
             {
                 wb_game = TRUE;
 
-                ResizeWindow(atol(resolutions[current_resolution]),
-                             atol(resolutions[current_resolution] + 4));
+                ResizeWindow(atoi(resolutions[current_resolution]),
+                             atoi(resolutions[current_resolution] + 4));
 
                 RedrawButton(&actual_menu->Button[16],
                              actual_menu->Button[0].Color);
